Added dotProduct() and normalized() methods to vectors
